requirements.txt
python-dotenv==1.0.0
pyotp>=2.9.0
git+https://github.com/Kotak-Neo/Kotak-neo-api-v2.git@v2.0.1#egg=neo_api_client
